+++ cisco_sdwan/tasks/implementation/_delete.py
@@ -40,24 +40,9 @@ class TaskDelete(Task):
 
         if parsed_args.detach:
             try:
-                # Deactivate vSmart policy
-                reqs = self.policy_deactivate(api, log_context='deactivating vSmart policy')
-                if reqs:
-                    self.log_debug(f'Deactivate requests processed: {reqs}')
-                else:
-                    self.log_info('No vSmart policy deactivate needed')
-
-                # Detach templates
                 template_index = DeviceTemplateIndex.get_raise(api)
-                # vSmarts
-                reqs = self.template_detach(api, template_index.filtered_iter(DeviceTemplateIndex.is_vsmart,
-                                                                              DeviceTemplateIndex.is_attached),
-                                            log_context='template detaching vSmarts')
-                if reqs:
-                    self.log_debug(f'Detach requests processed: {reqs}')
-                else:
-                    self.log_info('No vSmart template detachments needed')
-                # WAN Edges
+
+                # Detach WAN Edge templates
                 reqs = self.template_detach(api, template_index.filtered_iter(DeviceTemplateIndex.is_not_vsmart,
                                                                               DeviceTemplateIndex.is_attached),
                                             log_context='template detaching WAN Edges')
@@ -66,14 +51,37 @@ class TaskDelete(Task):
                 else:
                     self.log_info('No WAN Edge template detachments needed')
 
+                # Deactivate vSmart policy
+                reqs = self.policy_deactivate(api, log_context='deactivating vSmart policy')
+                if reqs:
+                    self.log_debug(f'Deactivate requests processed: {reqs}')
+                else:
+                    self.log_info('No vSmart policy deactivate needed')
+
+                # Detach vSmart templates
+                reqs = self.template_detach(api, template_index.filtered_iter(DeviceTemplateIndex.is_vsmart,
+                                                                              DeviceTemplateIndex.is_attached),
+                                            log_context='template detaching vSmarts')
+                if reqs:
+                    self.log_debug(f'Detach requests processed: {reqs}')
+                else:
+                    self.log_info('No vSmart template detachments needed')
+
                 # Dissociate WAN Edge config-groups
                 if is_index_supported(ConfigGroupIndex, version=api.server_version):
-                    reqs = self.cfg_group_dissociate(api, ConfigGroupIndex.get_raise(api),
-                                                     log_context='config-group dissociating WAN Edges')
-                    if reqs:
-                        self.log_debug(f'Dissociate requests processed: {reqs}')
-                    else:
-                        self.log_info('No WAN Edge config-group dissociate needed')
+                    config_groups = ConfigGroupIndex.get_raise(api)
+
+                    diss_reqs = self.cfg_group_dissociate(api, config_groups,
+                                                          log_context='config-group dissociating WAN Edges')
+                    if diss_reqs:
+                        self.log_debug(f'Dissociate requests processed: {diss_reqs}')
+
+                    rule_reqs = self.cfg_group_rules_delete(api, config_groups)
+                    if rule_reqs:
+                        self.log_debug(f'Automated rule delete requests processed: {rule_reqs}')
+
+                    if not (diss_reqs + rule_reqs):
+                        self.log_info('No WAN Edge config-group dissociate or automated rule deletes needed')
 
             except (RestAPIException, WaitActionsException) as ex:
                 self.log_critical(f'Detach failed: {ex}')