+++ CHANGELOG.md
@@ -1,9 +1,9 @@
-Sastre 1.19 [August xx, 2022]
+Sastre 1.19 [August 16, 2022]
 ================================
 
 #### Enhancements:
 - [#145] 20.8 vManage support, included support for the following API endpoints:
-  - Config group, 
+  - Config-group, Config-group tag rules
   - SDWAN system, SDWAN service, SDWAN transport and SDWAN cli feature profiles and parcels
 - Attach vSmart task has new --activate option to activate the centralized policy after vSmart template is attached.
 - Detach vSmart task now automatically deactivates the active centralized policy (if any) before detaching vSmart templates.